+++ src/main/kotlin/de/osca/fama/digitaltwin/TwinMessageManager.kt
@@ -5,6 +5,7 @@ import de.osca.fama.generated.BuildConfig
 import de.osca.fama.logger.logger
 import de.osca.fama.settings.Settings
 import de.osca.fama.smarthomeadapter.SmartHomeAdapter
+import io.sentry.kotlin.SentryContext
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.cancelAndJoin
@@ -31,7 +32,7 @@ object TwinMessageManager {
     private val logger by logger()
     private val smartHomeAdapter = SmartHomeAdapter.getAdapter(Settings.SMART_HOME_TYPE)
     private var session: StompSession? = null
-    private val dispatcher: CoroutineContext = Dispatchers.IO
+    private val dispatcher: CoroutineContext = Dispatchers.IO + SentryContext()
 
     private val subscriptions = mutableListOf<Job>()
 
@@ -108,7 +109,7 @@ object TwinMessageManager {
     private fun checkVersions(
         message: DigitalTwinMessage<*>,
         schemaVersion: Int = 1,
-        messageVersion: Int = 1,
+        messageVersion: Int = 1
     ): Boolean = message.schemaVersion == schemaVersion && message.messageVersion == messageVersion
 }