+++ src/main/kotlin/de/osca/fama/settings/EnvVarMissingException.kt
@@ -1,10 +1,9 @@
 package de.osca.fama.settings
 
-class EnvVarMissingException(
-    key: String,
-) : IllegalArgumentException(
-    """
+class EnvVarMissingException(key: String) :
+    IllegalArgumentException(
+        """
         Variable $key is missing. 
         Please read the documentation to ensure all required variables are set.
-    """.trimIndent()
-)
+    """
+    )